Update soft architecture

+++ src/rtda/Thread.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <stack>
 #include <memory>
 namespace rtda {
@@ -7,11 +8,11 @@ class Frame;
 class Stack;
 class Thread {
   private:
-  int pc;
+  int64_t pc;
   std::shared_ptr<Stack> stack;
   public:
   Thread();
-  int getPC() {return pc;}
+  int64_t getPC() {return pc;}
   void setPC(int pc) {this->pc = pc;}
   void pushFrame(std::shared_ptr<Frame> frame);
   std::shared_ptr<Frame> popFrame();
